refactor: LPのFAQコンポーネントを分離
